Fixed opensearch.rss query in browsers (#419)
A typical browser will do a GET request with text/html at the start of the accept header. The API incorrectly was setup to accept all content types, but in this case it can't render txt/html.
The solution is to only accept application/rss+xml,application/xml and application/xhtml+xml. Search.kml requests suffered from the same problem.